addressing review, part I
