+++ neuralmonkey/attention/__init__.py
@@ -1,0 +1,1 @@
+from .feed_forward import Attention
