+++ tests/bpe.ini
@@ -78,7 +78,7 @@ encoder_layers=2
 input_sequence=<encoder_input>
 
 [attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 state_size=8
 name="attention_sentence_encoder"
 encoder=<encoder>
