+++ tests/captioning.ini
@@ -54,7 +54,7 @@ network_type="vgg_16"
 attention_layer="vgg_16/conv5/conv5_3"
 
 [attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 state_size=10
 name="attention_sentence_encoder"
 encoder=<imagenet>
