+++ tests/factored.ini
@@ -72,7 +72,7 @@ data_ids=["source", "tags"]
 vocabularies=[<surface_source_vocabulary>, <tag_vocabulary>]
 
 [attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="attention_sentence_encoder"
 encoder=<encoder>
