+++ tests/hier-multiattention.ini
@@ -69,13 +69,13 @@ max_size=30
 overwrite=True
 
 [enc_attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="enc_attention"
 state_size=3
 encoder=<encoder>
 
 [img_attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="img_attention"
 state_size=2
 encoder=<imagenet>
