+++ tests/post-edit.ini
@@ -45,7 +45,7 @@ name="src_encoder"
 vocabulary=<source_vocabulary>
 
 [src_attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="attention_source_encoder"
 encoder=<src_encoder>
 
@@ -58,7 +58,7 @@ rnn_cell="LSTM"
 name="trans_encoder"
 
 [trans_attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="attention_trans_encoder"
 encoder=<trans_encoder>
