+++ tests/small.ini
@@ -61,7 +61,7 @@ data_id="source"
 vocabulary=<encoder_vocabulary>
 
 [attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="attention_sentence_encoder"
 encoder=<encoder>
