+++ tests/small_sent_cnn.ini
@@ -70,7 +70,7 @@ filters=[(1,13), (2,13), (3,13)]
 segment_size=5
 
 [attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="attention_sentence_encoder"
 encoder=<encoder>
