+++ tests/str.ini
@@ -63,7 +63,7 @@ convolutions=[(3, 3, None), (3, 3, 2)]
 fully_connected=[4,2]
 
 [attention]
-class=attention.base_attention.Attention
+class=attention.Attention
 name="attention_sentence_encoder"
 encoder=<encoder>
